feat(validation): add validation to routes

+++ app/Http/Controllers/UserController.php
@@ -4,11 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\UserResource;
 use App\Services\UserService;
+use Arr;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rules\Password;
 use Mockery\Exception;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Translation\Exception\NotFoundResourceException;
+use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
 {
@@ -24,12 +27,26 @@ class UserController extends Controller
     }
     public function getUserById(string $userId): JsonResponse
     {
-        // TODO: Validate the request
+            $validated = Validator::make(
+                ['userId' => $userId],
+                ['userId' => 'required|string|uuid'],
+            );
+
+            if($validated->fails()) {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'validation error',
+                    'errors' => $validated->messages()
+                ], 422);
+            }
 
         try {
             $user = UserService::findUserById($userId);
         } catch (NotFoundResourceException $exception) {
-            return response()->json(['errors' => $exception->getMessage()], $exception->getCode());
+            return response()->json([
+                'status' => false,
+                'errors' => $exception->getMessage()
+            ], $exception->getCode());
         }
 
         return (new UserResource($user))
@@ -39,14 +56,42 @@ class UserController extends Controller
     }
     public function createUser(Request $request): JsonResponse
     {
-        // TODO: Validate the request
-
         $payload = $request->only(["first_name", "last_name", "username", "date_of_birth", "email", "password", "phone_number"]);
+
+        $validated = Validator::make($payload, [
+            'first_name' => 'required|string|min:3|max:30',
+            'last_name' => 'required|string|min:3|max:30',
+            'username' => 'required|string|min:3|max:15|regex:/^\S*$/u|unique:users,username',
+            'date_of_birth' => 'required|date',
+            'email' => 'required|email|unique:users,email',
+            'password' => [
+                Password::min(8)   // must be at least 8 characters in length
+                    ->mixedCase()       // must contain mixed case
+                    ->numbers()         // must contain at least one digit
+                    ->symbols()         // must contain a special character
+                    ->uncompromised(),  // must not be a known compromised password
+            ],
+            'phone_number' => [
+                'required',
+                'regex:/^\+(9[976]\d|8[987530]\d|6[987]\d|5[90]\d|42\d|3[875]\d|2[98654321]\d|9[8543210]|8[6421]|6[6543210]|5[87654321]|4[987654310]|3[9643210]|2[70]|7|1)\d{1,14}$/'
+            ],
+        ]);
+
+        if($validated->fails()) {
+            return response()->json([
+                'status' => false,
+                'message' => 'validation error',
+                'errors' => $validated->messages()
+            ], 422);
+        }
 
         try {
             $model = UserService::createUser($payload);
         } catch (Exception $exception) {
-            return response()->json(["errors" => $exception->getMessage()], $exception->getCode());
+            return response()->json([
+                'status' => false,
+                'errors' => $exception->getMessage()
+            ], $exception->getCode());
         }
 
         return response()
@@ -55,25 +100,60 @@ class UserController extends Controller
     }
     public function updateUser(string $userId, Request $request): JsonResponse
     {
-        // TODO: Validate the request
-
         $payload = $request->only(["first_name", "last_name", "username", "date_of_birth", "email", "phone_number"]);
+
+        $validated = Validator::make(array_merge($payload, ["userId" => $userId]), [
+            'userId' => 'required|string|uuid',
+            'first_name' => 'string|min:3|max:30',
+            'last_name' => 'string|min:3|max:30',
+            'username' => 'string|min:3|max:15|regex:/^\S*$/u|unique:users,username,' . $userId,
+            'date_of_birth' => 'date',
+            'email' => 'email|unique:users,email,' . $userId,
+            'phone_number' => [
+                'regex:/^\+(9[976]\d|8[987530]\d|6[987]\d|5[90]\d|42\d|3[875]\d|2[98654321]\d|9[8543210]|8[6421]|6[6543210]|5[87654321]|4[987654310]|3[9643210]|2[70]|7|1)\d{1,14}$/'
+            ],
+        ]);
+
+        if($validated->fails()) {
+            return response()->json([
+                'status' => false,
+                'message' => 'validation error',
+                'errors' => $validated->messages()
+            ], 422);
+        }
 
         try {
             $model = UserService::updateUser($userId, $payload);
         } catch (NotFoundResourceException|Exception $exception) {
-            return response()->json(["errors" => $exception->getMessage()], $exception->getCode());
+            return response()->json([
+                'status' => false,
+                "errors" => $exception->getMessage()
+            ], $exception->getCode());
         }
 
         return response()->json(["data" => $model], Response::HTTP_OK);
     }
     public function deleteUser(string $userId): JsonResponse
     {
-        // TODO: Validate the request
+        $validated = Validator::make([$userId], [
+            'userId' => 'required|string|uuid',
+        ]);
+
+        if($validated->fails()) {
+            return response()->json([
+                'status' => false,
+                'message' => 'validation error',
+                'errors' => $validated->messages()
+            ], 422);
+        }
+
         try {
             UserService::deleteUser($userId);
         } catch (NotFoundResourceException|Exception $exception) {
-            return response()->json(["errors" => $exception->getMessage()], $exception->getCode());
+            return response()->json([
+                'status' => false,
+                'errors' => $exception->getMessage()
+            ], $exception->getCode());
         }
         return response()->json([], Response::HTTP_NO_CONTENT);
     }
